refactor: change styles for css modules
