+++ src/components/Header/Header.module.css
@@ -1,0 +1,26 @@
+.header {
+    background-color: #ffffff;
+    height: 80px;
+    justify-content: center;
+}
+
+.container {
+    display: flex;
+    width: 100%;
+    height: 100%;
+    max-width: 1680px;
+    gap: 30px;
+    margin: 0 auto;
+    align-items: center;
+    max-width: 1700px;
+    margin-left: auto;
+    margin-right: auto;
+    padding-left: 20px;
+    padding-right: 20px;
+    font-size: 1.5rem;
+    justify-content: flex-end;
+}
+
+.container:active {
+    color: #ff4500;
+}
