+++ src/components/components/CardItem/CardItem.module.css
@@ -1,0 +1,37 @@
+.item {
+    height: auto;
+    min-width: 371px;
+    max-width: 371px;
+    padding: 20px;
+    border-radius: 20px;
+    background-color: #fff;
+    box-shadow: 4px 4px 25px rgba(0, 0, 0, 0.05);
+    list-style-type: none;
+}
+
+.header {
+    width: 100%;
+    max-width: 361px;
+    height: 282px;
+}
+
+.image {
+    width: 100%;
+    height: 282px;
+    border-radius: 20px;
+}
+
+.content {
+    width: 100%;
+    margin-top: 15px;
+}
+
+.content span {
+    font-weight: bold;
+}
+
+.info {
+    font-size: 22px;
+    line-height: 26px;
+    color: rgba(28, 28, 28, 0.4);
+}
