+++ src/components/components/CardList/CardList.module.css
@@ -1,0 +1,8 @@
+.cardList {
+    margin-top: 20px;
+    display: flex;
+    column-gap: 1.3vw;
+    flex-wrap: wrap;
+    row-gap: 25px;
+    justify-content: center;
+}
