+++ src/pages/AboutPage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Header from "../components/Header";
+import Header from "../components/Header/Header";
 import styled from "styled-components";
 
 const StyledAboutPage = styled.div`
