+++ src/pages/FormsPage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Header from "../components/Header";
+import Header from "../components/Header/Header";
 import styled from "styled-components";
 import FormsSection from "../components/forms/FormsSection";
