+++ src/pages/HomePage/HomePage.module.css
@@ -1,0 +1,19 @@
+.homePage {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    height: 100%;
+    max-width: 1680px;
+    gap: 30px;
+    margin: 0 auto;
+    align-items: center;
+    max-width: 1700px;
+    margin-top: 25px;
+    justify-content: center;
+}
+
+.homePageLoading {
+    display: flex;
+    justify-content: center;
+    font-size: 2rem;
+}
