+++ src/tests/Header.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it } from "vitest";
 import { render, screen } from "@testing-library/react";
-import Header from "../components/Header";
+import Header from "../components/Header/Header";
 import React from "react";
 import { BrowserRouter } from "react-router-dom";
